// final/site/src/types/song.ts
export type Song = {
  track_id: string;
  track_name: string;
  artists: string;
  album_name: string;
  track_genre: string;
  cluster: string;
  x: number;
  y: number;
  energy: number;
  danceability: number;
  valence: number;
  tempo: number;
  acousticness: number;
  instrumentalness: number;
};